refactored everything into webpack, still crashing
