+++ basic-electron-react-boilerplate/package.json
@@ -15,7 +15,7 @@
     "Willy Ovalle <https://github.com/wovalle>"
   ],
   "license": "MIT",
-  "main": "main.js",
+  "main": "oldMain.js",
   "scripts": {
     "prod": "webpack --config webpack.build.config.js && electron --noDevServer .",
     "dev": "webpack-dev-server --hot --host 0.0.0.0 --port 4000 --config=./webpack.dev.config.js",
@@ -39,5 +39,9 @@
     "style-loader": "^0.19.0",
     "webpack": "^3.6.0",
     "webpack-dev-server": "^2.4.5"
+  },
+  "dependencies": {
+    "react-modal": "^3.14.2",
+    "react-router-dom": "^5.2.0"
   }
 }
