+++ basic-electron-react-boilerplate/src/data/config.json
