+++ basic-electron-react-boilerplate/src/index.js
@@ -1,6 +1,17 @@
 import React from 'react';
-import { render } from 'react-dom';
+import ReactDOM from 'react-dom';
+import './index.css';
 import App from './containers/App';
+import { BrowserRouter } from 'react-router-dom';
 
-// Now we can render our application into it
-render( <App />, document.getElementById('app') );
+ReactDOM.render((
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+  ), document.getElementById('root')
+);
+
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
